introduce IsomorphicMarkdown component
This component can be used to render either server-side-rendered markdown content that was previously serialised via next-mdx-remote or render a simple string of markdown text present on the client.

This is primarily necessary for client-side updates of content (replying to threads and editing posts) - the initial render uses SSR markdown, then subsequent renders use whatever markdown state that is updated when its parent re-renders.