rewritten the api fetcher functions to be more readable and ergonomic (#483)
* rewritten the api fetcher functions to be more readable and ergonomic

There's now a new fetcher, just called `api` which is isomorphic. It's much simpler in structure and easier to read. The way options are handled is much simpler, with just the schema and query params unioned with RequestInit.

Context is removed from the general API and instead `apiSSP` takes context as a required parameter.

Exported a request builder function for more specialised use-cases, such as the OAuth logins.

All other return values and APIs remain the same. There will be some future changes to the ergonomics of the API though, such as nicer SSP error handling.

* updated search swr to use non-keyed query params

* updated notifications swr and mutations to use non-keyed query params

* update post editor to use just the regular `api` function

* update tags input to use just the regular `api` function

* updated thread list view swr to use non-keyed query params

* improved apiSSP to be more ergonomic with errors and documented everything in fetcher.ts

see docstring comments for explanations

* added SSR support to thread list

* remove unused console.error